u32 -> usize in day1

+++ problems/src/problem1/mod.rs
@@ -1,7 +1,7 @@
 use failure::{Error, format_err};
 use utils::{split_by_lines, result, ProblemResult, RetOne};
 
-fn fuel_req(mass: u32) -> u32 {
+fn fuel_req(mass: usize) -> usize {
     let a = mass / 3;
     if a >= 2 {
         a - 2
@@ -10,7 +10,7 @@ fn fuel_req(mass: u32) -> u32 {
     }
 }
 
-fn first_star(input: &[u32]) -> ProblemResult<u32> {
+fn first_star(input: &[usize]) -> ProblemResult<usize> {
     let mut fuel = 0;
     for mass in input {
         fuel += fuel_req(*mass);
@@ -18,7 +18,7 @@ fn first_star(input: &[u32]) -> ProblemResult<u32> {
     Ok(fuel)
 }
 
-fn second_star(input: &[u32]) -> ProblemResult<u32> {
+fn second_star(input: &[usize]) -> ProblemResult<usize> {
     let mut fuel = 0;
 
     for mass in input {
@@ -32,10 +32,10 @@ fn second_star(input: &[u32]) -> ProblemResult<u32> {
     Ok(fuel)
 }
 
-pub(crate) fn solve() -> Result<RetOne<u32>, Error> {
+pub(crate) fn solve() -> Result<RetOne<usize>, Error> {
     let input_raw = include_str!("./input");
-    let input: Vec<u32> = split_by_lines(input_raw, &|e: &str|
-        e.parse::<u32>().or_else(|_| Err(format_err!("Failed to parse input"))))?;
+    let input: Vec<usize> = split_by_lines(input_raw, &|e: &str|
+        e.parse::<usize>().or_else(|_| Err(format_err!("Failed to parse input"))))?;
 
     Ok(result(first_star(&input), second_star(&input)))
 }
